Re-organise test dirs
