+++ .phpunit.cache/test-results
@@ -1,0 +1,1 @@
+{"version":1,"defects":{"PhpUnitSplitter\\Tests\\PhpUnitSplitterTest::testSplitter":5},"times":{"PhpUnitSplitter\\Tests\\PhpUnitSplitterTest::testSplitter":0.001}}
