+++ composer.json
@@ -12,15 +12,18 @@
   "require": {
     "php": "^8.1",
     "ext-simplexml": "*",
-    "phpunit/phpunit": "^9.6",
     "symfony/console": "^6.3"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9.6"
   },
   "autoload": {
     "psr-4": {"PhpUnitSplitter\\": "src/"}
   },
   "autoload-dev": {
     "psr-4": {
-      "PhpUnitSplitter\\Tests\\": "tests/"
+      "PhpUnitSplitter\\Tests\\": "tests/src",
+      "PhpUnitSplitter\\Tests\\Fixtures\\": "tests/fixtures/src"
     }
   },
   "config": {
