+++ scripts/generate-fixtures.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-./vendor/bin/phpunit --list-tests-xml=tests/fixtures/tests.xml tests/fixtures/Test
-./vendor/bin/phpunit tests/fixtures/Test --cache-result --cache-result-file=tests/fixtures/.phpunit.result.cache
+./vendor/bin/phpunit --list-tests-xml=tests/fixtures/tests.xml tests/fixtures/src
+./vendor/bin/phpunit tests/fixtures/src --cache-result --cache-result-file=tests/fixtures/.phpunit.result.cache
