+++ tests/fixtures/.phpunit.result.cache
@@ -1,1 +1,1 @@
-{"version":1,"defects":[],"times":{"PhpUnitSplitter\\Tests\\fixtures\\Test\\FastTestsTest::testOne":0.012,"PhpUnitSplitter\\Tests\\fixtures\\Test\\FastTestsTest::testTwo":0.02,"PhpUnitSplitter\\Tests\\fixtures\\Test\\FastTestsTest::testThree":0.03,"PhpUnitSplitter\\Tests\\fixtures\\Test\\FastTestsTest::testFour":0.04,"PhpUnitSplitter\\Tests\\fixtures\\Test\\FastTestsTest::testFive":0.05,"PhpUnitSplitter\\Tests\\fixtures\\Test\\SlowTestsTest::testOne":0.1,"PhpUnitSplitter\\Tests\\fixtures\\Test\\SlowTestsTest::testTwo":0.2,"PhpUnitSplitter\\Tests\\fixtures\\Test\\SlowTestsTest::testThree":0.3,"PhpUnitSplitter\\Tests\\fixtures\\Test\\SlowTestsTest::testFour":0.4,"PhpUnitSplitter\\Tests\\fixtures\\Test\\SlowTestsTest::testFive":0.5,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set #0":1,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set #1":0.111,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set #2":0.445,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set #3":0.222,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set #4":0.334,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set #5":0.667,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set \"one\"":0.111,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set \"two\"":0.445,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set \"three\"":0.222,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set \"four\"":0.334,"PhpUnitSplitter\\Tests\\fixtures\\Test\\ProviderTest::testProvider with data set \"five\"":0.667}}
+{"version":1,"defects":[],"times":{"PhpUnitSplitter\\Tests\\Fixtures\\FastTestsTest::testOne":0.011,"PhpUnitSplitter\\Tests\\Fixtures\\FastTestsTest::testTwo":0.02,"PhpUnitSplitter\\Tests\\Fixtures\\FastTestsTest::testThree":0.03,"PhpUnitSplitter\\Tests\\Fixtures\\FastTestsTest::testFour":0.04,"PhpUnitSplitter\\Tests\\Fixtures\\FastTestsTest::testFive":0.05,"PhpUnitSplitter\\Tests\\Fixtures\\ProviderTest::testProvider with data set \"one\"":0.111,"PhpUnitSplitter\\Tests\\Fixtures\\ProviderTest::testProvider with data set \"two\"":0.445,"PhpUnitSplitter\\Tests\\Fixtures\\ProviderTest::testProvider with data set \"three\"":0.222,"PhpUnitSplitter\\Tests\\Fixtures\\ProviderTest::testProvider with data set \"four\"":0.334,"PhpUnitSplitter\\Tests\\Fixtures\\ProviderTest::testProvider with data set \"five\"":0.667,"PhpUnitSplitter\\Tests\\Fixtures\\SlowTestsTest::testOne":0.1,"PhpUnitSplitter\\Tests\\Fixtures\\SlowTestsTest::testTwo":0.2,"PhpUnitSplitter\\Tests\\Fixtures\\SlowTestsTest::testThree":0.3,"PhpUnitSplitter\\Tests\\Fixtures\\SlowTestsTest::testFour":0.4,"PhpUnitSplitter\\Tests\\Fixtures\\SlowTestsTest::testFive":0.5}}
