+++ tests/fixtures/tests.xml
@@ -1,20 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <tests>
- <testCaseClass name="PhpUnitSplitter\Tests\fixtures\Test\FastTestsTest">
+ <testCaseClass name="PhpUnitSplitter\Tests\Fixtures\FastTestsTest">
   <testCaseMethod name="testOne" groups="default"/>
   <testCaseMethod name="testTwo" groups="default"/>
   <testCaseMethod name="testThree" groups="default"/>
   <testCaseMethod name="testFour" groups="default"/>
   <testCaseMethod name="testFive" groups="default"/>
  </testCaseClass>
- <testCaseClass name="PhpUnitSplitter\Tests\fixtures\Test\ProviderTest">
+ <testCaseClass name="PhpUnitSplitter\Tests\Fixtures\ProviderTest">
   <testCaseMethod name="testProvider" groups="default" dataSet="&quot;one&quot;"/>
   <testCaseMethod name="testProvider" groups="default" dataSet="&quot;two&quot;"/>
   <testCaseMethod name="testProvider" groups="default" dataSet="&quot;three&quot;"/>
   <testCaseMethod name="testProvider" groups="default" dataSet="&quot;four&quot;"/>
   <testCaseMethod name="testProvider" groups="default" dataSet="&quot;five&quot;"/>
  </testCaseClass>
- <testCaseClass name="PhpUnitSplitter\Tests\fixtures\Test\SlowTestsTest">
+ <testCaseClass name="PhpUnitSplitter\Tests\Fixtures\SlowTestsTest">
   <testCaseMethod name="testOne" groups="default"/>
   <testCaseMethod name="testTwo" groups="default"/>
   <testCaseMethod name="testThree" groups="default"/>
